Rename project from "demo" to "untitled" and add basic security configuration.
