+++ src/main/java/org/example/untitled/UntitledApplication.java
@@ -1,13 +1,13 @@
-package org.example.demo;
+package org.example.untitled;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DemoApplication {
+public class UntitledApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(DemoApplication.class, args);
+        SpringApplication.run(UntitledApplication.class, args);
     }
 
 }
