+++ src/main/java/org/example/untitled/config/SecurityConfig.java
@@ -1,0 +1,16 @@
+package org.example.untitled.config;
+
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+
+class SecurityConfig {
+
+    SecurityFilterChain securityFilterChain(HttpSecurity http) {
+        http.authorizeHttpRequests(auth -> auth
+                .requestMatchers("/", "/home").permitAll()
+                .anyRequest().permitAll()
+        );
+
+        return http.build();
+    }
+}
