+++ src/test/java/org/example/untitled/UntitledApplicationTests.java
@@ -1,4 +1,4 @@
-package org.example.demo;
+package org.example.untitled;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
